rename index ts to client page
