+++ app/[...filename]/page.tsx
@@ -1,4 +1,4 @@
-import Home from ".";
+import Home from "./client-page";
 import client from "../../tina/__generated__/client";
 
 export async function generateStaticParams() {
